requirements.txt
python-telegram-bot==12.2.0
django-environ==0.4.5
